synchro iced

+++ scripts/src/iced-0.4.0/iced/filter.py
@@ -155,8 +155,7 @@ def _filter_low_sum(X, percentage=0.02, remove_all_zeros_loci=False,
         x = X_sum[int(m * percentage)]
     else:
         num_noninteracting_loci = sum(X_sum == 0)
-        x = X_sum[int(len(X_sum[X_sum > 0]) * percentage) + num_noninteracting_loci]
-
+        x = X_sum[len(X_sum[X_sum > 0]) * percentage + num_noninteracting_loci]
 
     X_sum = np.array(X.sum(axis=0)).flatten()
